Move clientcmd into clusterctl since it's only needed by the CLI.
