+++ cmd/clusterctl/cmd/delete_cluster.go
@@ -21,10 +21,10 @@ import (
 
 	"k8s.io/api/core/v1"
 	tcmd "k8s.io/client-go/tools/clientcmd"
+	"sigs.k8s.io/cluster-api/cmd/clusterctl/clientcmd"
 	"sigs.k8s.io/cluster-api/cmd/clusterctl/clusterdeployer"
 	"sigs.k8s.io/cluster-api/cmd/clusterctl/clusterdeployer/bootstrap/minikube"
 	"sigs.k8s.io/cluster-api/cmd/clusterctl/providercomponents"
-	"sigs.k8s.io/cluster-api/pkg/clientcmd"
 
 	"github.com/golang/glog"
 	"github.com/spf13/cobra"
