+++ cmd/clusterctl/cmd/validate_cluster.go
@@ -22,8 +22,8 @@ import (
 
 	"github.com/spf13/cobra"
 	tcmd "k8s.io/client-go/tools/clientcmd"
+	"sigs.k8s.io/cluster-api/cmd/clusterctl/clientcmd"
 	"sigs.k8s.io/cluster-api/cmd/clusterctl/validation"
-	"sigs.k8s.io/cluster-api/pkg/clientcmd"
 )
 
 type ValidateClusterOptions struct {
